Adding Web3 functionality to the website
Added functionality to interact with the smart contract using webBrowser and metamask.

For the website to interact with the blockchain network, use the web3.js library.

change in the truffle-config.js file is so that you can connect to ganache's local blockchain node.

Abi file needed to interact with the contract from webpage. Can access it from build/contract/FroggyContract.json