CreateGen0 Frog web button
Added functionality to the website to call the createFrogGen0 when create frog is pressed in the web ui.

Click listener waiting for the createFrogButton to be pressed, which calls the function createFrog() in the index.js file.

createFrog uses the instance of the contract to call createFrogGen0() in the contract.

A prompt to sign tx using Metamask will pop up and Tx history should show in Ganache is connected properly.