Adding IERC721 approval functions
The implementation of ERC721 didn't implement all the functionality of a ERC721 token.

This change will implement the new IERC721 interface provided, which adds approval functionality to the contract.

The approval functionality will not be used in this contract, it was implemented so that I can get experience of how to implement a function from the "EIP-721: Non-Fungible Token Standard".

These functions must be used for the function to qualify as a ERC721 token.