modifcation of the IERC721 transfer functions
Changed some of the functions from external to public so that they could be accessed from within the contract code.

Refactored safeTransferFrom() using a try/catch block, this was inspired by the implementation in the ERC721 standard contract.

This implementation of safeTranferFrom does work and was tested with the Dummy contract I created (included in this commit)