Refactoring safeTransferFrom
In the last commit, created a safeTransfer, which will be used by the safeTransferFrom() functions created in this commit.

Allows the user to safely send the token to another address, with confidence that the recipient address isn't a contract address that can't handle ERC721 tokens.

(contracts that can't implement ERC721 tokens, result in tokens being sent to that address, but the contract can do nothing with it since it can't call any functions in the token's smart contract)