Loading Portfolio
Added asyncronous function to the for loop to get the frog details of each individual frog and load it to the correct frog elements.

Without the async function, the for loop would loop through all the numberOfFrogs, creating them all on the sceen but not render a single one, as the programme would still be waiting for the getFrogDetails() to be called. getFrogDetails() would then only call after the loop condition finished, which is when i is larger than the length of the frogs array, and i would stay the same. This resulted in all frogs created getting frogDetails of a frog that didnt exist.

Using the await, don't continue the next frog until we get the current details