Breed Frogs
Select 2 frogs and create a breed a new frog.

The breed function is already created in the smart contract. So this is to call the function from the front end.

Main bulk of code is int he loadPortfolio.js file. Code listens for button press, and when there are 2 frogs selected, the smart contract breed function called, and the page is reloaded after 10 seconds.