Refactored the removeOffer() in Markeplace
This code will remove the tokenId from list of tokenId's, by swapping the last tokenId in the list with the token that's being removed.

This doesn't remove the tokenId from the mapping, but we don't care as _offerExists() function makes it seem the offer never existed when called as a check in setOffer().