bug: Loading portfolio
When loading the frogs from the portfolio, I would get the amount of frogs a account owns as n, then loop through from 0-n and get the details of frogs from 0-n. This initially worked, but when a frog was transferred from one user to the other and it wasn’t the last frog created, this method of displaying the portfolio doesn’t work.

There are two methods to solve this scenario. Both involve looping through the total supply of frogs, and if the account was the ownerOf a frog, add it’s id to an array of frogs it owned. This could be done on the contract or on the website.

Doing it on the website is bad, as each iteration would need to call the ownerOf() function. Whereas if this was done in the contract as a function call, as it’s not changing the blockchain, it’s just a view function and won’t cost anything. Then we can simply query the array from the contract with one call.

Added the tokensOfOwner function which returns an array of all frogs an address owns