Moving keys to private file
When uploading to GitHub you don’t want to share your private keys which could be accidentally done as the truffle-config.js needs your private key to sign the tx.
The solution to this is to move your private keys to a separate file and make truffle-config read the values from this file. Adding the file containing your private keys, to .gitignore will make sure that git ignores that file when making commits