Updating the compiler
When using pragma solidity >=0.4.22 <0.9.0, for some reason the the explicit change to a payable address in FroggyMarket.sol would throw errors when I tried to compile.

Using "truffle compile --list", I took a version of solidity and updated the versions in my solidity files.

The new compiler requested that I need to add "override" to the functions which were "overridden" from the interfaces.

All tests then passed.